requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

# RAG
langchain==0.2.9
langchain-community==0.2.7
chromadb==0.5.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1

# STT / TTS / OCR / vision
openai-whisper==20231117
gTTS==2.5.3
pydub==0.25.1
pytesseract==0.3.10
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84

# utilities
python-dotenv==1.0.1
